refactor: Separate profile-related loaders/hooks from core feature
