+++ src/features/core/api/loaders.ts
@@ -1,14 +1,7 @@
 import { QueryClient } from "@tanstack/react-query";
 import { currentUserQuery } from "../utils/auth";
-import { allPostsQuery } from "../../explore/api/queries";
 import { ActionFunctionArgs, ParamParseKey, Params } from "react-router-dom";
-import {
-  followQuery,
-  likeQuery,
-  postQuery,
-  userPostsQuery,
-  usersByUsernameQuery,
-} from "./queries";
+import { followQuery, likeQuery, postQuery } from "./queries";
 
 const Paths = {
   postDetail: "/p/:postId",
@@ -21,28 +14,8 @@ interface PhotoLoaderArgs extends ActionFunctionArgs {
   params: Params<ParamParseKey<typeof Paths.postDetail>>;
 }
 
-interface ProfileLoaderArgs extends ActionFunctionArgs {
-  params: Params<ParamParseKey<typeof Paths.profileDetail>>;
-}
-
-interface ProfilePostsLoaderArgs extends ActionFunctionArgs {
-  params: Params<ParamParseKey<typeof Paths.profilePostsDetail>>;
-}
-
 export const appLoader = (queryClient: QueryClient) => async () => {
   const query = currentUserQuery();
-  return await queryClient
-    .ensureQueryData(query)
-    .then((data) => {
-      return data;
-    })
-    .catch(() => {
-      return null;
-    });
-};
-
-export const exploreLoader = (queryClient: QueryClient) => async () => {
-  const query = allPostsQuery();
   return await queryClient
     .ensureQueryData(query)
     .then((data) => {
@@ -93,64 +66,4 @@ export const postDetailsLoader =
       : null;
 
     return { post, like, follow };
-  };
-
-export const profileLoader =
-  (queryClient: QueryClient) =>
-  async ({ params }: ProfileLoaderArgs) => {
-    if (!params.username) {
-      return { user: null, follow: null };
-    }
-
-    const user = await queryClient
-      .ensureQueryData(usersByUsernameQuery(params.username))
-      .then((data) => {
-        return data[0];
-      })
-      .catch(() => {
-        return null;
-      });
-    const follow =
-      user ?
-        await queryClient
-          .ensureQueryData(followQuery(user._id))
-          .then((data) => {
-            return data;
-          })
-          .catch(() => {
-            return null;
-          })
-      : null;
-
-    return { user, follow };
-  };
-
-export const profilePostsLoader =
-  (queryClient: QueryClient) =>
-  async ({ params }: ProfilePostsLoaderArgs) => {
-    if (!params.username) {
-      return { user: null, posts: [] };
-    }
-
-    const user = await queryClient
-      .ensureQueryData(usersByUsernameQuery(params.username))
-      .then((data) => {
-        return data[0];
-      })
-      .catch(() => {
-        return null;
-      });
-    const posts =
-      user ?
-        await queryClient
-          .ensureQueryData(userPostsQuery(user._id))
-          .then((data) => {
-            return data;
-          })
-          .catch(() => {
-            return null;
-          })
-      : null;
-
-    return { user, posts };
   };
