+++ src/features/posts/hooks/usePostQuery.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
-import { postQuery } from "../api/queries";
-import ObjectId from "../types/objectId";
-import Post from "../types/post";
+import { postQuery } from "../../core/api/queries";
+import ObjectId from "../../core/types/objectId";
+import Post from "../../core/types/post";
 
 export default function usePostQuery(postId: ObjectId, initialData?: Post) {
   return useQuery({ ...postQuery(postId), initialData, enabled: !!postId });
